requirements.txt
pynput == 1.4.2
selenium == 3.141.0
pytest == 4.4.0
browsermob-proxy == 0.8.0
allure-pytest == 2.6.2
allure-python-commons == 2.6.2
mysql-connector == 2.2.9
